+++ src/examples/scala/com/decisionbrain/cplex/cp/HouseBuilding.scala
@@ -1,12 +1,10 @@
 /*
  * Source file provided under Apache License, Version 2.0, January 2004,
  * http://www.apache.org/licenses/
- * (c) Copyright DecisionBrain SAS 2016,2017
+ * (c) Copyright DecisionBrain SAS 2016,2018
  */
 
 package com.decisionbrain.cplex.cp
-
-import ilog.concert.IloIntervalVar
 
 import com.decisionbrain.cplex.cp.CpModel._