+++ src/main/scala/com/decisionbrain/cplex/Addable.scala
@@ -1,7 +1,7 @@
 /*
  * Source file provided under Apache License, Version 2.0, January 2004,
  * http://www.apache.org/licenses/
- * (c) Copyright DecisionBrain SAS 2016,2017
+ * (c) Copyright DecisionBrain SAS 2016,2018
  */
 
 package com.decisionbrain.cplex
@@ -9,9 +9,6 @@ package com.decisionbrain.cplex
 import ilog.concert.IloAddable
 
 
-/**
-  * Created by dgodard on 10/02/2017.
-  */
 trait Addable {
 
   /**