+++ src/main/scala/com/decisionbrain/cplex/cp/IntExpr.scala
@@ -1,13 +1,12 @@
 /*
  * Source file provided under Apache License, Version 2.0, January 2004,
  * http://www.apache.org/licenses/
- * (c) Copyright DecisionBrain SAS 2016,2017
+ * (c) Copyright DecisionBrain SAS 2016,2018
  */
 
 package com.decisionbrain.cplex.cp
 
-import ilog.concert.{IloIntExpr, IloNumExpr}
-import ilog.cp.IloCP
+import ilog.concert.IloIntExpr
 
 /**
   * Created by dgodard on 11/02/2017.