+++ src/main/scala/com/decisionbrain/cplex/cp/IntVar.scala
@@ -1,14 +1,13 @@
 /*
  * Source file provided under Apache License, Version 2.0, January 2004,
  * http://www.apache.org/licenses/
- * (c) Copyright DecisionBrain SAS 2016,2017
+ * (c) Copyright DecisionBrain SAS 2016,2018
  */
 
 package com.decisionbrain.cplex.cp
 
 import com.decisionbrain.cplex.Addable
 import ilog.concert.{IloAddable, IloIntVar, IloNumVar, IloNumVarType}
-import ilog.cp.IloCP
 
 /**
   * Class for numeric variables.