+++ src/main/scala/com/decisionbrain/cplex/cp/NumToNumStepFunction.scala
@@ -1,7 +1,7 @@
 /*
  * Source file provided under Apache License, Version 2.0, January 2004,
  * http://www.apache.org/licenses/
- * (c) Copyright DecisionBrain SAS 2016,2017
+ * (c) Copyright DecisionBrain SAS 2016,2018
  */
 
 package com.decisionbrain.cplex.cp
@@ -12,12 +12,33 @@ import ilog.concert.cppimpl.IloConcertUtils
 import ilog.cp.IloCP
 
 /**
+  * Iterator on step function.
+  *
+  * @param f is the step function
+  * @param model is the constraint programming model
+  */
+class NumToNumStepFunctionIterator(f: NumToNumStepFunction)(implicit model: CpModel) extends Iterator[NumStep] {
+
+  val cursor = model.numToNumStepFunctionCursor(f, f.getDefinitionIntervalMin())
+
+  override def hasNext: Boolean = cursor.ok()
+
+  override def next(): NumStep = {
+    val segment = NumStep(cursor.getSegmentMin,
+      cursor.getSegmentMax,
+      cursor.getValue)
+    cursor.next
+    segment
+  }
+}
+
+/**
   * Constructor of class NumToNumStepFunction.
   *
   * @param f is a CPLEX step function
   * @param model is the constraint programming model
   */
-class NumToNumStepFunction(f: IloNumToNumStepFunction)(implicit model: CpModel) {
+class NumToNumStepFunction(f: IloNumToNumStepFunction)(implicit model: CpModel) extends Iterable[NumStep] {
 
   /**
     * Returns the CPLEX step function
@@ -287,20 +308,20 @@ class NumToNumStepFunction(f: IloNumToNumStepFunction)(implicit model: CpModel) 
     var prevX = xmin
     var prevV = .0
     val builder = new StringBuilder
-    builder.append("NumToNuMStepFunction: [").append(xmin).append(" .. ").append(xmax).append(") {")
-    val cursor = model.numToNumStepFunctionCursor(this, this.getDefinitionIntervalMin)
-    while (cursor.ok()) {
-      val x = cursor.getSegmentMin
-      val v = cursor.getValue
+    builder.append("NumToNumStepFunction: [").append(xmin).append(" .. ").append(xmax).append(") {")
+    for (s <- this) {
+      val x = s.start
+      val v = s.value
       if (prevX < x) builder.append("[").append(prevX).append(" .. ").append(x).append(") -> ").append(prevV).append(" ; ")
       prevX = x
       prevV = v
-      cursor.next()
     }
     builder.append("[").append(prevX).append(" .. ").append(xmax).append(") -> ").append(prevV).append(" ; ")
     builder.append("}")
     builder.toString
   }
+
+  override def iterator: Iterator[NumStep] = new NumToNumStepFunctionIterator(this)
 }
 
 object NumToNumStepFunction {