+++ src/main/scala/com/decisionbrain/cplex/mp/MpModel.scala
@@ -1,14 +1,14 @@
 /*
  * Source file provided under Apache License, Version 2.0, January 2004,
  * http://www.apache.org/licenses/
- * (c) Copyright DecisionBrain SAS 2016,2017
+ * (c) Copyright DecisionBrain SAS 2016,2018
  */
 
 package com.decisionbrain.cplex.mp
 
 import com.decisionbrain.cplex.Addable
 import com.decisionbrain.cplex.mp.NumExpr.{LinearIntExpr, LinearNumExpr}
-import ilog.cplex.IloCplex
+import ilog.opl.IloCplex
 
 
 /**