+++ src/main/scala/com/decisionbrain/cplex/mp/RangeConstraint.scala
@@ -1,13 +1,12 @@
 /*
  * Source file provided under Apache License, Version 2.0, January 2004,
  * http://www.apache.org/licenses/
- * (c) Copyright DecisionBrain SAS 2016,2017
+ * (c) Copyright DecisionBrain SAS 2016,2018
  */
 
 package com.decisionbrain.cplex.mp
 
 import ilog.concert.IloRange
-import ilog.cplex.IloCplex
 
 /**
   * Created by dgodard on 09/02/2017.