+++ src/test/scala/com/decisionbrain/cplex/cp/TalentTest.scala
@@ -6,20 +6,25 @@
 
 package com.decisionbrain.cplex.cp
 
+import ilog.cp.IloCP
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
 import org.scalatest.{FunSuite, Matchers}
 
 @RunWith(classOf[JUnitRunner])
-class SchedSquareTest extends FunSuite with Matchers {
+class TalentTest extends FunSuite with Matchers {
 
   val epsilon = 1e-6
 
-  test("SchedSquare") {
-    val model = SchedSquare.build()
+  test("Talent") {
+    val model = Talent.build("data/rehearsal.data")
+
+    model.cp.setParameter(IloCP.DoubleParam.TimeLimit, 10.0)
+
     val status = model.solve()
 
     status should equal(true)
+    model.getObjectiveValue() should equal(17.0 +- epsilon)
 
     model.end()
   }