+++ src/test/scala/com/decisionbrain/cplex/cp/TruckFleetTest.scala
@@ -1,22 +1,15 @@
 /*
  * Source file provided under Apache License, Version 2.0, January 2004,
  * http://www.apache.org/licenses/
- * (c) Copyright DecisionBrain SAS 2016,2017
+ * (c) Copyright DecisionBrain SAS 2016,2018
  */
 
 package com.decisionbrain.cplex.cp
 
-import com.decisionbrain.cplex.cp.Color._
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
 import org.scalatest.{FunSuite, Matchers}
 
-/**
-  * Created by dgodard on 11/02/2017.
-  */
-/**
-  * Created by dgodard on 07/02/2017.
-  */
 @RunWith(classOf[JUnitRunner])
 class TruckFleetTest extends FunSuite with Matchers {
 
@@ -24,7 +17,7 @@ class TruckFleetTest extends FunSuite with Matchers {
 
   test("TruckFleet") {
     val model = TruckFleet.build()
-    val status = model.solve(timeLimit=5, logPeriod=3000)
+    val status = model.solve(timeLimit=30, solutionLimit=12, logPeriod=3000)
 
     status should equal(true)